requirements.txt
flask
python-dotenv
line-bot-sdk
openai==1.14.2
gunicorn

firebase-admin==6.0.1
Flask>=2.2.0
Werkzeug>=2.2.0
gunicorn==20.1.0

gspread==5.12.4
google-auth==2.29.0